promise形式のコード作成
